some restructuring of debug libs; updates #1303
